Moved my project list into it's own container component to keep index.js page cleaner
